ttml: remove bogus casts
ptrdiff_t is the result type for a substraction of pointers, not the
operands type for the substraction. The cast made no sense.

To convert pointers to signed integers, intptr_t shall be used. But
there is no point in casting here: equality and inequality are always
well-defined for valid and/or NULL pointers.

+++ modules/demux/ttml.c
@@ -432,7 +432,7 @@ int OpenDemux( vlc_object_t* p_this )
 
     /* Simplified probing. Valid TTML must have a namespace declaration */
     const char *psz_tt = strnstr( psz_xml, "tt ", i_xml );
-    if( !psz_tt || (ptrdiff_t)psz_tt == (ptrdiff_t)psz_xml ||
+    if( !psz_tt || psz_tt == psz_xml ||
         (psz_tt[-1] != ':' && psz_tt[-1] != '<') )
     {
         free( psz_alloc );
